+++ spinetoolbox/project_items/gdx_export/__init__.py
@@ -10,8 +10,22 @@
 ######################################################################################################################
 
 """
-Init file for spine_io package. Intentionally empty.
+Gdx export project item plugin.
 
-:author: P. Vennström (VTT)
-:date:   1.6.2019
+:author: A. Soininen (VTT)
+:date:   25.9.2019
 """
+
+from project_items.gdx_export.gdx_export import GdxExport
+from project_items.gdx_export.gdx_export_icon import GdxExportIcon
+from project_items.gdx_export.widgets.add_gdx_export_widget import AddGdxExportWidget
+from project_items.gdx_export.widgets.gdx_export_properties import GdxExportProperties
+
+item_rank = 5
+item_category = "Data Exporters"
+item_type = "Gdx Export"
+item_icon = ":/icons/project_item_icons/file-export-solid.svg"
+item_maker = GdxExport
+icon_maker = GdxExportIcon
+properties_widget_maker = GdxExportProperties
+add_form_maker = AddGdxExportWidget